Support elemMatch for fast index filters
An `elemMatch` is not that much different from a regular filter. The catch is that you have to check multiple elements in an array for a subtree match. But since that subtree doesn't skip any property steps, the search is rather limited.

Beyond that, the index behaves the same as non-elemMatch indexes; if any element of elemMatch matches the nestedQuery then the node gets added to the index for the value where it resolved to. This means that one node may end up in multiple buckets. For example:

```
node = { a: [{b: 1}, {b: 2}] }
```

For a filter like `a.elemMatch.b` the index would create two buckets (value is `1` and `2`) and put the node in each bucket. Normally a filter only resolves to one value, so ends up in one bucket.

The cache key for an `elemMatch` is the same as for regular queries; the array positions are simply replaced by `elemMatch`, which runs no risk of collisions by definition of how `elemMatch` works.

At lookup time (when actually resolving a query) the process is the same as regular queries; The filter cache `type/a.elemMatch.b/$eq` contains value buckets and the `elemMatch` query will, ultimately, target a specific value. Get that value, look it up in the reverse cache, and you'll get the bucket that matches all nodes with at least one eleeent matching the filter.